+++ HISTORY.md
@@ -4,9 +4,10 @@ Release History
 1.1.0 (2023-12-08)
 ------------------
 
-- Add api-key limits 
+- Add api-key limits
 - Add insight schemas API
-- Add filebox API
+- Add filebox API 
+- Add enrichment tasks API
 
 1.0.8 (2023-12-06)
 ------------------